Adjusting Submit Button
I made some changes to the submit button on the entry frame that make the program more usable - the entry fields clearing when the button is pressed and a confirmation message to show that the job has submitted. Code to clear the entry has been moved to its own method.