Error handling for Distance and Customer Name entry
I fixed the boundaries of the distance slider by setting the start value to 1 instead of 0. I also added onto the submit method so that it will not submit if there is no customer name entered, meaning the user must enter something there for the submission of the job to go through, stopping them from creating a job with no customer.