Adding Edit Functionality
I added a button to the display frame that allows you to edit the job that you are currently on. This brings up the entry frame, but with all the details of the job in the checkboxes/scale/entries, and the user can save changes to that job specifically by pressing a different button that grids in place of the submit job button. The adding new job functionality still works as a seperate thing.