Adding the Summary Frame and Final Testing
I added the frame into the main code that would show all of the jobs in a summary frame, and added the methods that would update this frame to show the relevant details, as well as add functionality to the buttons that would switch between the frames. Then I tested the program by running through the testing plan one final time to ensure that nothing had been broken (the video unfortunately has to be low quality because github does not allow large files).